+++ lib/src/channel/axeptio.dart
@@ -55,6 +55,22 @@ class AxeptioSdk {
     );
   }
 
+  Future<Map<int, bool>> getVendorConsents() {
+    return AxeptioSdkPlatform.instance.getVendorConsents();
+  }
+
+  Future<List<int>> getConsentedVendors() {
+    return AxeptioSdkPlatform.instance.getConsentedVendors();
+  }
+
+  Future<List<int>> getRefusedVendors() {
+    return AxeptioSdkPlatform.instance.getRefusedVendors();
+  }
+
+  Future<bool> isVendorConsented(int vendorId) {
+    return AxeptioSdkPlatform.instance.isVendorConsented(vendorId);
+  }
+
   addEventListerner(AxeptioEventListener listener) {
     AxeptioSdkPlatform.instance.addEventListener(listener);
   }